Backpropagation implementation of  mergemax, mergeadd and mergeavg ops (#343)
* libnd4j: first step of merge_max implementation


* libnd4j fixed typos


* libnd4j some corrections for mergeMaxBp


* libnd4j some minor corrections


* libnd4j test added for mergemax_bp


* libnd4j fixed several problems tests added, check with gradCheck


* libnd4j remove duplicated tests


* libnd4j split implementation of transforms ops into separate file implementation


* libnd4j code clean up, added mergeavg_bp and mergeadd_bp, need testing


* libnd4j merge master, fixed typos and added tests


* libnd4j some minor fixes


* libnd4j added helper for mergeAddBp operation, this permits to skip nullify


* libnd4j file renaming changes and cuda some corrections, need some additional corrections


* libnd4j some additional corrections for merge ops


* libnd4j more corrections per request for cuda more proper usage
