+++ libnd4j/include/ops/declarable/generic/transforms/merge_max.cpp
@@ -33,7 +33,7 @@ OP_IMPL(mergemax, -1, 1, false) {
         
     auto output = OUTPUT_VARIABLE(0);
 
-    std::vector<NDArray*> inArrs(block.width());
+    std::vector<const NDArray*> inArrs(block.width());
     
     for(int i = 0; i < block.width(); ++i)
         inArrs[i] = INPUT_VARIABLE(i);
@@ -42,13 +42,53 @@ OP_IMPL(mergemax, -1, 1, false) {
 
     return Status::OK();
 }
-
 DECLARE_SYN(MergeMax, mergemax);
 
     DECLARE_TYPES(mergemax) {
         getOpDescriptor()
                 ->setAllowedInputTypes(sd::DataType::ANY)
                 ->setAllowedOutputTypes(sd::DataType::ANY);
+    }
+
+
+    CUSTOM_OP_IMPL(mergemax_bp, 2, 1, false, 0, 0) {
+
+        auto inSize = block.width();
+
+        REQUIRE_OK(this->validateInputDimensionsMatch(block));
+
+        std::vector<const NDArray*> inArrs(inSize);
+        std::vector<NDArray*> outArrs(inSize - 1);
+
+        for (int i = 0; i < inSize; ++i)
+            inArrs[i] = INPUT_VARIABLE(i);
+
+        for (int i = 0; i < (inSize - 1); ++i) {
+            outArrs[i] = OUTPUT_NULLIFIED(i);
+        }
+
+        helpers::mergeMaxBp(block.launchContext(), inArrs, outArrs);
+
+        return Status::OK();
+    }
+
+    DECLARE_TYPES(mergemax_bp) {
+        getOpDescriptor()
+            ->setAllowedInputTypes(sd::DataType::ANY)
+            ->setAllowedOutputTypes(sd::DataType::ANY);
+    }
+    DECLARE_SHAPE_FN(mergemax_bp) {
+
+        const int numOfInArrs = block.width() - 1;
+
+        auto shapeList = SHAPELIST();
+        
+        for (int e = 0; e < numOfInArrs; e++) {
+            auto inShape = inputShape->at(e);
+             shapeList->push_back(ConstantShapeHelper::getInstance()->createShapeInfo(ShapeDescriptor(ArrayOptions::dataType(inShape), shape::order(inShape), shape::shapeOf(inShape), shape::rank(inShape))));
+        }
+
+        return shapeList;
     }
 
 }
