+++ libnd4j/include/ops/declarable/headers/transforms.h
@@ -64,6 +64,7 @@ namespace sd {
 
         #if NOT_EXCLUDED(OP_mergemax)
         DECLARE_OP(mergemax, -1, 1, false);
+        DECLARE_CUSTOM_OP(mergemax_bp, 2, 1, false, 0, 0);
         #endif
         /*
          * Complete tensor with max indices merged from all input tensors list
@@ -78,10 +79,12 @@ namespace sd {
 
         #if NOT_EXCLUDED(OP_mergeadd)
         DECLARE_OP(mergeadd, -1, 1, false);
+        DECLARE_CUSTOM_OP(mergeadd_bp, 2, 1, false, 0, 0);
         #endif
 
         #if NOT_EXCLUDED(OP_mergeavg)
         DECLARE_OP(mergeavg, -1, 1, false);
+        DECLARE_CUSTOM_OP(mergeavg_bp, 2, 1, false, 0, 0);
         #endif
 
         #if NOT_EXCLUDED(OP_scatter_update)
